+++ de.metas.material/dispo-service/src/main/sql/postgresql/system/5462220_sys_gh523-metasfresh_AD_Process_MD_Candidate_Request_PP_Order_fix_classname.sql
@@ -1,0 +1,6 @@
+
+-- 2017-05-03T14:59:14.905
+-- I forgot to set the DICTIONARY_ID_COMMENTS System Configurator
+UPDATE AD_Process SET Classname='de.metas.material.dispo.service.process.MD_Candidate_Request_PP_Order',Updated=TO_TIMESTAMP('2017-05-03 14:59:14','YYYY-MM-DD HH24:MI:SS'),UpdatedBy=100 WHERE AD_Process_ID=540784
+;
+